backend/api/requirements.txt
asgiref==3.2.10
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
lorem==0.1.1
Markdown==3.2.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
